requirements.txt
gradio
google-generativeai
pandas
openpyxl